+++ next-docs/public/examples/button/FullWidth.tsx
@@ -1,0 +1,9 @@
+import { Button } from '@heathmont/moon-components';
+
+const Example = () => (
+  <Button fullWidth variant="primary">
+    Action
+  </Button>
+);
+
+export default Example;