+++ next-docs/public/examples/checkbox/AriaLabel.tsx
@@ -1,0 +1,7 @@
+import { Checkbox } from '@heathmont/moon-core';
+
+const Example = () => (
+  <Checkbox ariaLabel="Description of the checkbox's purpose." />
+);
+
+export default Example;