+++ next-docs/public/examples/chip/States.tsx
@@ -1,0 +1,10 @@
+import { Chip } from '@heathmont/moon-components';
+
+const Example = () => (
+  <div className="flex justify-around items-center w-full">
+    <Chip isActive={true}>Active</Chip>
+    <Chip>Inactive</Chip>
+  </div>
+);
+
+export default Example;