+++ next-docs/public/examples/select/select/Default.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import { Select } from '@heathmont/moon-components';
+
+const Example = () => (
+  <Select>
+    <option value="">Select an option</option>
+    <option value="never-gonna">Never gonna</option>
+    <option value="give-you-up">give you up</option>
+    <option value="let-you-down">let you down</option>
+  </Select>
+);
+
+export default Example;