+++ next-docs/public/examples/switch/Theme.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import { useState } from 'react';
+
+import { Switch } from '@heathmont/moon-components';
+
+const Example = () => {
+  const [state, setState] = useState(true);
+
+  return (
+    <>
+      <Switch
+        checked={state}
+        onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+          setState(e.target.checked)
+        }
+        colorScheme
+      />
+    </>
+  );
+};
+
+export default Example;